get rid of empty string logic in addressPage and move to prefill
