+++ src/applications/health-care-supply-reordering/config/prefill-transformer.js
@@ -4,6 +4,31 @@ import {
   isTerritory,
 } from '../utils/addresses';
 
+const cleanEmptyAddressFields = address => {
+  if (!address || typeof address !== 'object') {
+    return address;
+  }
+
+  const fieldsToClean = [
+    'street',
+    'street2',
+    'city',
+    'state',
+    'postalCode',
+    'country',
+  ];
+
+  const cleanedAddress = { ...address };
+
+  fieldsToClean.forEach(fieldName => {
+    if (cleanedAddress[fieldName] === '') {
+      cleanedAddress[fieldName] = undefined;
+    }
+  });
+
+  return cleanedAddress;
+};
+
 export default function prefillTransformer(pages, formData, metadata) {
   const newFormData = formData;
 
@@ -26,6 +51,13 @@ export default function prefillTransformer(pages, formData, metadata) {
           ? newFormData[addressType].street2
           : undefined;
     }
+  }
+
+  // Clean empty strings from temporaryAddress
+  if (newFormData.temporaryAddress) {
+    newFormData.temporaryAddress = cleanEmptyAddressFields(
+      newFormData.temporaryAddress,
+    );
   }
 
   return {
