+++ src/applications/health-care-supply-reordering/pages/addressPage.js
@@ -1,6 +1,3 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { setData } from 'platform/forms-system/src/js/actions';
 import {
   addressSchema,
   addressUI,
@@ -39,54 +36,9 @@ const optionalAddressRequired = {
   country: () => false,
 };
 
-// Clean fields that are empty strings and set to undefined so it does not flag addressPattern
-const cleanEmptyAddressFields = address => {
-  const fieldsToClean = [
-    'street',
-    'street2',
-    'city',
-    'state',
-    'postalCode',
-    'country',
-  ];
-
-  const cleanedAddress = { ...address };
-
-  fieldsToClean.forEach(fieldName => {
-    if (cleanedAddress[fieldName] === '') {
-      cleanedAddress[fieldName] = undefined;
-    }
-  });
-
-  return cleanedAddress;
-};
-
-const AddressPageCleanupWrapper = ({ children }) => {
-  const dispatch = useDispatch();
-  const formData = useSelector(state => state.form?.data);
-
-  useEffect(() => {
-    const temporaryAddress = formData?.temporaryAddress;
-    if (temporaryAddress && typeof temporaryAddress === 'object') {
-      const cleanedAddress = cleanEmptyAddressFields(temporaryAddress);
-      if (cleanedAddress !== temporaryAddress) {
-        dispatch(
-          setData({
-            ...formData,
-            temporaryAddress: cleanedAddress,
-          }),
-        );
-      }
-    }
-  }, []);
-
-  return <>{children}</>;
-};
-
 /** @type {PageSchema} */
 export default {
   uiSchema: {
-    'ui:description': AddressPageCleanupWrapper,
     [permanentAddressField]: {
       ...addressUiWithDlWrappedFields(),
       'ui:title': 'Permanent address',
@@ -112,32 +64,6 @@ export default {
           formData['view:currentAddress'] !== 'temporaryAddress',
         viewComponent: AddressViewField,
       },
-      'ui:validations': [
-        (errors, field, formData) => {
-          const temporaryAddress = formData?.temporaryAddress;
-          if (temporaryAddress && typeof temporaryAddress === 'object') {
-            const fieldsToCheck = [
-              'street',
-              'street2',
-              'city',
-              'state',
-              'postalCode',
-              'country',
-            ];
-
-            fieldsToCheck.forEach(fieldName => {
-              if (temporaryAddress[fieldName] === '') {
-                Object.assign(formData, {
-                  temporaryAddress: {
-                    ...formData.temporaryAddress,
-                    [fieldName]: undefined,
-                  },
-                });
-              }
-            });
-          }
-        },
-      ],
     },
     [vetEmailField]: {
       ...emailUI(),
